Create Container_CarbonExtractor
First "draft" of an air filter using the Carbon Extractor as a base. By default, the Carbon Extractor is placed in the Specialized Electrics tech node, which is way higher than USI-LS's small recycler, which goes into the Survivability node. Since this is a low efficiency recycler, I think it should go there unless CTT is installed, then it can go into Recycling (which is where CTT puts it). 
I am not 100% sure if "@TechRequired:NEEDS[!CommunityTechTree] = survivability" is the right way to do that or not.